requirements.txt
pandas
langchain
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
groq
python-dotenv
langgraph